+++ struct_excelrow.go
