+++ struct_infobarnum.go
