+++ struct_ocioconfig.go
