+++ struct_partner.go
