+++ struct_publishkey.go
