+++ struct_review.go
