+++ struct_rocketchat.go
