+++ struct_searchop.go
