+++ struct_setellite.go
