+++ struct_setting.go
