+++ struct_stage.go
