+++ struct_status.go
