+++ stuct_camera.go
